boring::ec

Struct EcGroup

pub struct EcGroup(/* private fields */);

Describes the curve

A curve can be of the named curve type. These curves can be discovered using openssl binary openssl ecparam -list_curves. Other operations are available in the wiki. These named curves are available in the Nid module.

Curves can also be generated using prime field parameters or a binary field.

Prime fields use the formula y^2 mod p = x^3 + ax + b mod p. Binary fields use the formula y^2 + xy = x^3 + ax^2 + b. Named curves have assured security. To prevent accidental vulnerabilities, they should be preferred.

Implementations

impl EcGroup

pub fn from_curve_name(nid: Nid) -> Result<EcGroup, ErrorStack>

Returns the group of a standard named curve.

This corresponds to EC_GROUP_new.