boring::ec

Struct EcGroupRef

pub struct EcGroupRef(/* private fields */);

A borrowed reference to a EcGroup.

Implementations

impl EcGroupRef

pub fn components_gfp( &self, p: &mut BigNumRef, a: &mut BigNumRef, b: &mut BigNumRef, ctx: &mut BigNumContextRef, ) -> Result<(), ErrorStack>

Places the components of a curve over a prime field in the provided BigNums. The components make up the formula y^2 mod p = x^3 + ax + b mod p.

OpenSSL documentation available at EC_GROUP_get_curve_GFp

pub fn cofactor( &self, cofactor: &mut BigNumRef, ctx: &mut BigNumContextRef, ) -> Result<(), ErrorStack>

Places the cofactor of the group in the provided BigNum.

This corresponds to EC_GROUP_get_cofactor.

pub fn degree(&self) -> u32

Returns the degree of the curve.

This corresponds to EC_GROUP_get_degree.

pub fn order_bits(&self) -> u32

Returns the number of bits in the group order.

This corresponds to EC_GROUP_order_bits.

pub fn generator(&self) -> &EcPointRef

Returns the generator for the given curve as a EcPoint.

This corresponds to EC_GROUP_get0_generator.

pub fn order( &self, order: &mut BigNumRef, ctx: &mut BigNumContextRef, ) -> Result<(), ErrorStack>

Places the order of the curve in the provided BigNum.

This corresponds to EC_GROUP_get_order.

pub fn set_asn1_flag(&mut self, flag: Asn1Flag)

Sets the flag determining if the group corresponds to a named curve or must be explicitly parameterized.

This defaults to EXPLICIT_CURVE in OpenSSL 1.0.1 and 1.0.2, but NAMED_CURVE in OpenSSL 1.1.0.

pub fn curve_name(&self) -> Option<Nid>

Returns the name of the curve, if a name is associated.

This corresponds to EC_GROUP_get_curve_name.