boring::pkey

Struct PKeyRef

pub struct PKeyRef<T>(/* private fields */);

Reference to PKey.

Implementations

impl<T> PKeyRef<T>

pub fn rsa(&self) -> Result<Rsa<T>, ErrorStack>

Returns a copy of the internal RSA key.

This corresponds to EVP_PKEY_get1_RSA.

pub fn dsa(&self) -> Result<Dsa<T>, ErrorStack>

Returns a copy of the internal DSA key.

This corresponds to EVP_PKEY_get1_DSA.

pub fn dh(&self) -> Result<Dh<T>, ErrorStack>

Returns a copy of the internal DH key.

This corresponds to EVP_PKEY_get1_DH.

pub fn ec_key(&self) -> Result<EcKey<T>, ErrorStack>

Returns a copy of the internal elliptic curve key.

This corresponds to EVP_PKEY_get1_EC_KEY.

pub fn id(&self) -> Id

Returns the Id that represents the type of this key.

This corresponds to EVP_PKEY_id.

pub fn size(&self) -> usize

Returns the maximum size of a signature in bytes.

This corresponds to EVP_PKEY_size.

impl<T> PKeyRef<T>
where T: HasPublic,

pub fn public_key_to_pem(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a PEM-encoded SubjectPublicKeyInfo structure.

The output will have a header of -----BEGIN PUBLIC KEY-----.

This corresponds to PEM_write_bio_PUBKEY.

pub fn public_key_to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the public key into a DER-encoded SubjectPublicKeyInfo structure.

This corresponds to i2d_PUBKEY.

pub fn bits(&self) -> u32

Returns the size of the key.

This corresponds to the bit length of the modulus of an RSA key, and the bit length of the group order for an elliptic curve key, for example.

pub fn public_eq<U>(&self, other: &PKeyRef<U>) -> bool
where U: HasPublic,

Compares the public component of this key with another.

pub fn raw_public_key_len(&self) -> Result<usize, ErrorStack>

Returns the length of the “raw” form of the public key. Only supported for certain key types.

This corresponds to EVP_PKEY_get_raw_public_key.

pub fn raw_public_key<'a>( &self, out: &'a mut [u8], ) -> Result<&'a [u8], ErrorStack>

Outputs a copy of the “raw” form of the public key. Only supported for certain key types.

Returns the used portion of out.

This corresponds to EVP_PKEY_get_raw_public_key.

impl<T> PKeyRef<T>
where T: HasPrivate,

pub fn private_key_to_pem_pkcs8(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded PKCS#8 PrivateKeyInfo structure.

The output will have a header of -----BEGIN PRIVATE KEY-----.

This corresponds to PEM_write_bio_PKCS8PrivateKey.

pub fn private_key_to_pem_pkcs8_passphrase( &self, cipher: Cipher, passphrase: &[u8], ) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a PEM-encoded PKCS#8 EncryptedPrivateKeyInfo structure.

The output will have a header of -----BEGIN ENCRYPTED PRIVATE KEY-----.

This corresponds to PEM_write_bio_PKCS8PrivateKey.

pub fn private_key_to_der(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a DER-encoded key type specific format.

This corresponds to i2d_PrivateKey.

pub fn private_key_to_der_pkcs8(&self) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a DER-encoded PKCS#8 PrivateKeyInfo structure.

This corresponds to i2d_PKCS8PrivateKey_bio.

pub fn private_key_to_der_pkcs8_passphrase( &self, cipher: Cipher, passphrase: &[u8], ) -> Result<Vec<u8>, ErrorStack>

Serializes the private key to a DER-encoded PKCS#8 EncryptedPrivateKeyInfo structure.

This corresponds to i2d_PKCS8PrivateKey_bio.

pub fn raw_private_key_len(&self) -> Result<usize, ErrorStack>

Returns the length of the “raw” form of the private key. Only supported for certain key types.

This corresponds to EVP_PKEY_get_raw_private_key.

pub fn raw_private_key<'a>( &self, out: &'a mut [u8], ) -> Result<&'a [u8], ErrorStack>

Outputs a copy of the “raw” form of the private key. Only supported for certain key types.

Returns the used portion of out.

This corresponds to EVP_PKEY_get_raw_private_key.