boring::ssl

Struct ExtensionType

pub struct ExtensionType(/* private fields */);

Extension types, to be used with ClientHello::get_extension.

WARNING: The current implementation of From is unsound, as it’s possible to create an ExtensionType that is not defined by the impl. From will be deprecated in favor of TryFrom in the next major bump of the library.

Implementations

impl ExtensionType

pub const SERVER_NAME: Self

pub const STATUS_REQUEST: Self

pub const EC_POINT_FORMATS: Self

pub const SIGNATURE_ALGORITHMS: Self

pub const SRTP: Self

pub const APPLICATION_LAYER_PROTOCOL_NEGOTIATION: Self

pub const PADDING: Self

pub const EXTENDED_MASTER_SECRET: Self

pub const QUIC_TRANSPORT_PARAMETERS_LEGACY: Self

pub const QUIC_TRANSPORT_PARAMETERS_STANDARD: Self

pub const CERT_COMPRESSION: Self

pub const SESSION_TICKET: Self

pub const SUPPORTED_GROUPS: Self

pub const PRE_SHARED_KEY: Self

pub const EARLY_DATA: Self

pub const SUPPORTED_VERSIONS: Self

pub const COOKIE: Self

pub const PSK_KEY_EXCHANGE_MODES: Self

pub const CERTIFICATE_AUTHORITIES: Self

pub const SIGNATURE_ALGORITHMS_CERT: Self

pub const KEY_SHARE: Self

pub const RENEGOTIATE: Self

pub const DELEGATED_CREDENTIAL: Self

pub const APPLICATION_SETTINGS: Self

pub const ENCRYPTED_CLIENT_HELLO: Self

pub const CERTIFICATE_TIMESTAMP: Self

pub const NEXT_PROTO_NEG: Self

pub const CHANNEL_ID: Self