boring::ssl

Struct Ssl

pub struct Ssl(/* private fields */);

The state of an SSL/TLS session.

Ssl objects are created from an SslContext, which provides configuration defaults. These defaults can be overridden on a per-Ssl basis, however.

Implementations

impl Ssl

pub fn new_ex_index<T>() -> Result<Index<Ssl, T>, ErrorStack>
where T: 'static + Sync + Send,

Returns a new extra data index.

Each invocation of this function is guaranteed to return a distinct index. These can be used to store data in the context that can be retrieved later by callbacks, for example.

This corresponds to SSL_get_ex_new_index.

pub fn new(ctx: &SslContextRef) -> Result<Ssl, ErrorStack>

Creates a new Ssl.

This corresponds to SSL_new.

pub fn setup_connect<S>(self, stream: S) -> MidHandshakeSslStream<S>
where S: Read + Write,

Initiates a client-side TLS handshake, returning a MidHandshakeSslStream.

This method is guaranteed to return without calling any callback defined in the internal Ssl or SslContext.

See SslStreamBuilder::setup_connect for more details.

Warning

BoringSSL’s default configuration is insecure. It is highly recommended to use SslConnector rather than Ssl directly, as it manages that configuration.

pub fn connect<S>(self, stream: S) -> Result<SslStream<S>, HandshakeError<S>>
where S: Read + Write,

Attempts a client-side TLS handshake.

This is a convenience method which combines Self::setup_connect and MidHandshakeSslStream::handshake.

Warning

OpenSSL’s default configuration is insecure. It is highly recommended to use SslConnector rather than Ssl directly, as it manages that configuration.

pub fn setup_accept<S>(self, stream: S) -> MidHandshakeSslStream<S>
where S: Read + Write,

Initiates a server-side TLS handshake.

This method is guaranteed to return without calling any callback defined in the internal Ssl or SslContext.

See SslStreamBuilder::setup_accept for more details.

Warning

BoringSSL’s default configuration is insecure. It is highly recommended to use SslAcceptor rather than Ssl directly, as it manages that configuration.

pub fn accept<S>(self, stream: S) -> Result<SslStream<S>, HandshakeError<S>>
where S: Read + Write,

Attempts a server-side TLS handshake.

This is a convenience method which combines Self::setup_accept and MidHandshakeSslStream::handshake.

Warning

OpenSSL’s default configuration is insecure. It is highly recommended to use SslAcceptor rather than Ssl directly, as it manages that configuration.