boring::ssl

Trait AsyncPrivateKeyMethod

pub trait AsyncPrivateKeyMethod:
    Send
    + Sync
    + 'static {
    // Required methods
    fn sign(
        &self,
        ssl: &mut SslRef,
        input: &[u8],
        signature_algorithm: SslSignatureAlgorithm,
        output: &mut [u8],
    ) -> Result<BoxPrivateKeyMethodFuture, AsyncPrivateKeyMethodError>;
    fn decrypt(
        &self,
        ssl: &mut SslRef,
        input: &[u8],
        output: &mut [u8],
    ) -> Result<BoxPrivateKeyMethodFuture, AsyncPrivateKeyMethodError>;
}

Describes async private key hooks. This is used to off-load signing operations to a custom, potentially asynchronous, backend. Metadata about the key such as the type and size are parsed out of the certificate.

See PrivateKeyMethod for the sync version of those hooks.

Required Methods

fn sign( &self, ssl: &mut SslRef, input: &[u8], signature_algorithm: SslSignatureAlgorithm, output: &mut [u8], ) -> Result<BoxPrivateKeyMethodFuture, AsyncPrivateKeyMethodError>

Signs the message input using the specified signature algorithm.

This method uses a function that returns a future whose output is itself a closure that will be passed ssl and output to finish writing the signature.

See PrivateKeyMethod::sign for the sync version of this method.

fn decrypt( &self, ssl: &mut SslRef, input: &[u8], output: &mut [u8], ) -> Result<BoxPrivateKeyMethodFuture, AsyncPrivateKeyMethodError>

Decrypts input.

This method uses a function that returns a future whose output is itself a closure that will be passed ssl and output to finish decrypting the input.

See PrivateKeyMethod::decrypt for the sync version of this method.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".