boring::ssl

Trait CertificateCompressor

pub trait CertificateCompressor:
    Send
    + Sync
    + 'static {
    const ALGORITHM: CertificateCompressionAlgorithm;
    const CAN_COMPRESS: bool;
    const CAN_DECOMPRESS: bool;

    // Provided methods
    fn compress<W>(&self, input: &[u8], output: &mut W) -> Result<()>
       where W: Write { ... }
    fn decompress<W>(&self, input: &[u8], output: &mut W) -> Result<()>
       where W: Write { ... }
}

Describes certificate compression algorithm. Implementation MUST implement transformation at least in one direction.

Required Associated Constants

const ALGORITHM: CertificateCompressionAlgorithm

An IANA assigned identifier of compression algorithm

const CAN_COMPRESS: bool

Indicates if compressor support compression

const CAN_DECOMPRESS: bool

Indicates if compressor support decompression

Provided Methods

fn compress<W>(&self, input: &[u8], output: &mut W) -> Result<()>
where W: Write,

Perform compression of input buffer and write compressed data to output.

fn decompress<W>(&self, input: &[u8], output: &mut W) -> Result<()>
where W: Write,

Perform decompression of input buffer and write compressed data to output.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".