boring::symm

Struct Cipher

pub struct Cipher(/* private fields */);

Represents a particular cipher algorithm.

See OpenSSL doc at EVP_EncryptInit for more information on each algorithms.

Implementations

impl Cipher

pub fn from_nid(nid: Nid) -> Option<Cipher>

Looks up the cipher for a certain nid.

This corresponds to EVP_get_cipherbynid.

pub fn aes_128_ecb() -> Cipher

pub fn aes_128_cbc() -> Cipher

pub fn aes_128_ctr() -> Cipher

pub fn aes_128_gcm() -> Cipher

pub fn aes_128_ofb() -> Cipher

pub fn aes_192_ecb() -> Cipher

pub fn aes_192_cbc() -> Cipher

pub fn aes_192_ctr() -> Cipher

pub fn aes_192_gcm() -> Cipher

pub fn aes_192_ofb() -> Cipher

pub fn aes_256_ecb() -> Cipher

pub fn aes_256_cbc() -> Cipher

pub fn aes_256_ctr() -> Cipher

pub fn aes_256_gcm() -> Cipher

pub fn aes_256_ofb() -> Cipher

pub fn des_cbc() -> Cipher

pub fn des_ecb() -> Cipher

pub fn des_ede3() -> Cipher

pub fn des_ede3_cbc() -> Cipher

pub fn rc4() -> Cipher

pub unsafe fn from_ptr(ptr: *const EVP_CIPHER) -> Cipher

Creates a Cipher from a raw pointer to its OpenSSL type.

Safety

The caller must ensure the pointer is valid for the 'static lifetime.

pub fn as_ptr(&self) -> *const EVP_CIPHER

pub fn key_len(&self) -> usize

Returns the length of keys used with this cipher.

pub fn iv_len(&self) -> Option<usize>

Returns the length of the IV used with this cipher, or None if the cipher does not use an IV.

pub fn block_size(&self) -> usize

Returns the block size of the cipher.

Note

Stream ciphers such as RC4 have a block size of 1.

pub fn nid(&self) -> Nid

Returns the cipher’s NID.

This corresponds to EVP_CIPHER_nid.