boring::x509

Struct X509ReqBuilder

pub struct X509ReqBuilder(/* private fields */);

A builder used to construct an X509Req.

Implementations

impl X509ReqBuilder

pub fn new() -> Result<X509ReqBuilder, ErrorStack>

Returns a builder for a certificate request.

This corresponds to X509_REQ_new.

pub fn set_version(&mut self, version: i32) -> Result<(), ErrorStack>

Set the numerical value of the version field.

This corresponds to X509_REQ_set_version.

pub fn set_subject_name( &mut self, subject_name: &X509NameRef, ) -> Result<(), ErrorStack>

Set the issuer name.

This corresponds to X509_REQ_set_subject_name.

pub fn set_pubkey<T>(&mut self, key: &PKeyRef<T>) -> Result<(), ErrorStack>
where T: HasPublic,

Set the public key.

This corresponds to X509_REQ_set_pubkey.

pub fn x509v3_context<'a>( &'a self, conf: Option<&'a ConfRef>, ) -> X509v3Context<'a>

Return an X509v3Context. This context object can be used to construct certain X509 extensions.

pub fn add_extensions( &mut self, extensions: &StackRef<X509Extension>, ) -> Result<(), ErrorStack>

Permits any number of extension fields to be added to the certificate.

pub fn sign<T>( &mut self, key: &PKeyRef<T>, hash: MessageDigest, ) -> Result<(), ErrorStack>
where T: HasPrivate,

Sign the request using a private key.

This corresponds to X509_REQ_sign.

pub fn build(self) -> X509Req

Returns the X509Req.