boring::x509

Struct X509VerifyError

pub struct X509VerifyError(/* private fields */);

Implementations

impl X509VerifyError

pub unsafe fn from_raw(err: c_int) -> X509VerifyResult

Creates an X509VerifyResult from a raw error number.

Safety

Some methods on X509VerifyError are not thread safe if the error number is invalid.

pub fn as_raw(&self) -> c_int

Return the integer representation of an X509VerifyError.

pub fn error_string(&self) -> &'static str

Return a human readable error string from the verification error.

Returns empty string if the message was not UTF-8.

This corresponds to X509_verify_cert_error_string.

impl X509VerifyError

pub const UNSPECIFIED: Self

pub const UNABLE_TO_GET_ISSUER_CERT: Self

pub const UNABLE_TO_GET_CRL: Self

pub const UNABLE_TO_DECRYPT_CERT_SIGNATURE: Self

pub const UNABLE_TO_DECRYPT_CRL_SIGNATURE: Self

pub const UNABLE_TO_DECODE_ISSUER_PUBLIC_KEY: Self

pub const CERT_SIGNATURE_FAILURE: Self

pub const CRL_SIGNATURE_FAILURE: Self

pub const CERT_NOT_YET_VALID: Self

pub const CERT_HAS_EXPIRED: Self

pub const CRL_NOT_YET_VALID: Self

pub const CRL_HAS_EXPIRED: Self

pub const ERROR_IN_CERT_NOT_BEFORE_FIELD: Self

pub const ERROR_IN_CERT_NOT_AFTER_FIELD: Self

pub const ERROR_IN_CRL_LAST_UPDATE_FIELD: Self

pub const ERROR_IN_CRL_NEXT_UPDATE_FIELD: Self

pub const OUT_OF_MEM: Self

pub const DEPTH_ZERO_SELF_SIGNED_CERT: Self

pub const SELF_SIGNED_CERT_IN_CHAIN: Self

pub const UNABLE_TO_GET_ISSUER_CERT_LOCALLY: Self

pub const UNABLE_TO_VERIFY_LEAF_SIGNATURE: Self

pub const CERT_CHAIN_TOO_LONG: Self

pub const CERT_REVOKED: Self

pub const INVALID_CA: Self

pub const PATH_LENGTH_EXCEEDED: Self

pub const INVALID_PURPOSE: Self

pub const CERT_UNTRUSTED: Self

pub const CERT_REJECTED: Self

pub const SUBJECT_ISSUER_MISMATCH: Self

pub const AKID_SKID_MISMATCH: Self

pub const AKID_ISSUER_SERIAL_MISMATCH: Self

pub const KEYUSAGE_NO_CERTSIGN: Self

pub const UNABLE_TO_GET_CRL_ISSUER: Self

pub const UNHANDLED_CRITICAL_EXTENSION: Self

pub const KEYUSAGE_NO_CRL_SIGN: Self

pub const UNHANDLED_CRITICAL_CRL_EXTENSION: Self

pub const INVALID_NON_CA: Self

pub const PROXY_PATH_LENGTH_EXCEEDED: Self

pub const KEYUSAGE_NO_DIGITAL_SIGNATURE: Self

pub const PROXY_CERTIFICATES_NOT_ALLOWED: Self

pub const INVALID_EXTENSION: Self

pub const INVALID_POLICY_EXTENSION: Self

pub const NO_EXPLICIT_POLICY: Self

pub const DIFFERENT_CRL_SCOPE: Self

pub const UNSUPPORTED_EXTENSION_FEATURE: Self

pub const UNNESTED_RESOURCE: Self

pub const PERMITTED_VIOLATION: Self

pub const EXCLUDED_VIOLATION: Self

pub const SUBTREE_MINMAX: Self

pub const APPLICATION_VERIFICATION: Self

pub const UNSUPPORTED_CONSTRAINT_TYPE: Self

pub const UNSUPPORTED_CONSTRAINT_SYNTAX: Self

pub const UNSUPPORTED_NAME_SYNTAX: Self

pub const CRL_PATH_VALIDATION_ERROR: Self

pub const HOSTNAME_MISMATCH: Self

pub const EMAIL_MISMATCH: Self

pub const IP_ADDRESS_MISMATCH: Self

pub const INVALID_CALL: Self

pub const STORE_LOOKUP: Self

pub const NAME_CONSTRAINTS_WITHOUT_SANS: Self