boring::x509::verify

Struct X509VerifyParamRef

pub struct X509VerifyParamRef(/* private fields */);

A borrowed reference to a X509VerifyParam.

Implementations

impl X509VerifyParamRef

pub fn set_flags(&mut self, flags: X509VerifyFlags)

Set verification flags.

This corresponds to X509_VERIFY_PARAM_set_flags.

pub fn clear_flags(&mut self, flags: X509VerifyFlags)

Clear verification flags.

This corresponds to X509_VERIFY_PARAM_clear_flags.

pub fn set_hostflags(&mut self, hostflags: X509CheckFlags)

Set the host flags.

This corresponds to X509_VERIFY_PARAM_set_hostflags.

pub fn flags(&self) -> X509VerifyFlags

Gets verification flags.

This corresponds to X509_VERIFY_PARAM_get_flags.

pub fn set_host(&mut self, host: &str) -> Result<(), ErrorStack>

Set the expected DNS hostname.

This corresponds to X509_VERIFY_PARAM_set1_host.

pub fn set_email(&mut self, email: &str) -> Result<(), ErrorStack>

Set the expected email address.

This corresponds to X509_VERIFY_PARAM_set1_email.

pub fn set_ip(&mut self, ip: IpAddr) -> Result<(), ErrorStack>

Set the expected IPv4 or IPv6 address.

This corresponds to X509_VERIFY_PARAM_set1_ip.

pub fn set_time(&mut self, time: time_t)

Set the verification time, where time is of type time_t, traditionaly defined as seconds since the epoch

This corresponds to X509_VERIFY_PARAM_set_time.

pub fn set_depth(&mut self, depth: c_int)

Set the verification depth

This corresponds to X509_VERIFY_PARAM_set_depth.

pub fn copy_from(&mut self, src: &Self) -> Result<(), ErrorStack>

Copies parameters from src.

If a parameter is unset in src, the existing value in `self`` is preserved.

This corresponds to X509_VERIFY_PARAM_set1.